+++ src/main/java/hekireki/sanjijiksong/global/security/dto/ReissueResponse.java
@@ -1,0 +1,8 @@
+package hekireki.sanjijiksong.global.security.dto;
+
+import jakarta.servlet.http.Cookie;
+
+public record ReissueResponse (
+        String newAccessToken,
+        Cookie refreshCookie
+){}
